ほぼすべてのIntをUByteに変更

+++ src/main/kotlin/Main.kt
@@ -1,7 +1,7 @@
 package io.github.jinnosukeKato
 
 fun main() {
-    val cpu = TD4("./src/main/resources/test2.bin", 0b1010)
+    val cpu = TD4("./src/main/resources/test2.bin", (0b1010).toUByte())
     for (i in 0..100) {
         val operation = cpu.fetch()
         cpu.execute(operation.first, operation.second)
